requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
openpyxl>=3.1.0
python-calamine>=0.2.0
pyarrow>=14.0.0
plotly>=5.18.0
xlsxwriter>=3.1.0
python-dateutil>=2.8.0
anthropic>=0.40.0
